restructure theme files
