+++ frontend/src/theme/common/buttonStyles.tsx
@@ -1,5 +1,5 @@
 import type { ComponentStyleConfig } from "@chakra-ui/theme";
-import textStyles from "./common/textStyles";
+import textStyles from "../textStyles";
 
 const Button: ComponentStyleConfig = {
   baseStyle: {
