--- frontend/src/theme/fields/textAreaStyles.tsx
@@ -1,41 +1,0 @@
-import type { ComponentStyleConfig } from "@chakra-ui/theme";
-import textStyles from "../common/textStyles";
-
-const Textarea: ComponentStyleConfig = {
-  sizes: {
-    sm: {
-      fontSize: "sm",
-    },
-    md: {
-      fontSize: "md",
-    },
-    lg: {
-      fontSize: "lg",
-    },
-  },
-  variants: {
-    default: {
-      border: "1px solid",
-      borderColor: "teal.100",
-      borderRadius: "4px",
-      color: "teal.300",
-      height: "36px",
-      _placeholder: {
-        color: "gray.600",
-      },
-      _hover: {
-        borderColor: "teal.400",
-      },
-      _focus: {
-        borderColor: "teal.400",
-      },
-      ...textStyles["body-medium"],
-    },
-  },
-  defaultProps: {
-    variant: "default",
-    size: "sm",
-  },
-};
-
-export default Textarea;
